+++ CHANGELOG.md
@@ -9,7 +9,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Display the name of the current functional test (#293)
 
 ### Changed
-- Convert more tests to nimut/testing-framework (#283, #288, #290, #291, #292)
+- Convert more tests to nimut/testing-framework (#283, #288, #290, #291, #292, #300)
 - Update the testing libraries (#275, #279)
 - Mark tests that do not contain any assertions (#277)